+++ @xen-orchestra/lite/src/stories/web-core/ui/radio-button-group/ui-radio-button-group.story.vue
@@ -1,29 +1,37 @@
 <template>
   <ComponentStory
-    v-slot="{ properties, settings }"
+    v-slot="{ properties }"
     :params="[
-      prop('accent').required().enum('brand', 'warning', 'danger').preset('brand').widget(),
+      prop('accent').required().enum('info', 'warning', 'danger').preset('info').widget(),
+      prop('label').str().widget().preset('Label'),
+      prop('info').str().widget().preset('message'),
       prop('vertical').bool().widget(),
-      prop('labels').required().preset(labels).widget(),
-      slot(),
-      slot('info'),
-      setting('defaultSlot').widget(text()).preset('Label'),
-      setting('info').widget(text()).preset('Lorem ipsum dolor sit amet, consectetur adipiscing elit.'),
+      slot().help('Meant to receive a list of checkboxes component'),
+      slot('label').help('Meant to receive a label UiLabel component or another component'),
+      slot('info').help('Meant to receive a message info or UiInfo component or another component'),
     ]"
   >
-    <UiRadioButtonGroup v-bind="properties" :labels="labels">
-      {{ settings.defaultSlot }}
-      <template v-if="settings.info" #info>{{ settings.info }}</template>
+    <UiRadioButtonGroup v-bind="properties">
+      <UiRadioButton
+        v-for="(label, index) in labels"
+        :key="index"
+        v-model="selectedRadio"
+        v-bind="properties"
+        :value="`radio-${index}`"
+      >
+        {{ label }}
+      </UiRadioButton>
     </UiRadioButtonGroup>
   </ComponentStory>
 </template>
 
 <script lang="ts" setup>
 import ComponentStory from '@/components/component-story/ComponentStory.vue'
-import { prop, setting, slot } from '@/libs/story/story-param'
-import { text } from '@/libs/story/story-widget'
+import { prop, slot } from '@/libs/story/story-param'
+import UiRadioButton from '@core/components/ui/radio-button/UiRadioButton.vue'
 import UiRadioButtonGroup from '@core/components/ui/radio-button-group/UiRadioButtonGroup.vue'
 import { ref } from 'vue'
+const selectedRadio = ref('')
 
 const labels = ref(['Label 1', 'Label 2', 'Label 3'])
 </script>
